+++ package.json
@@ -4,6 +4,9 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@chakra-ui/react": "^1.4.2",
+    "@emotion/react": "^11.1.5",
+    "@emotion/styled": "^11.1.5",
     "@material-ui/core": "^4.11.3",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
@@ -11,6 +14,8 @@
     "animate.css": "^4.1.1",
     "axios": "^0.21.1",
     "formik": "^2.2.6",
+    "framer-motion": "^4.1.2",
+    "public-ip": "^4.0.3",
     "react": "^16.14.0",
     "react-animated-text": "^0.1.4",
     "react-cookie": "^4.0.3",
